Separate .dictlike; make .util a directory
